1. Check G_loc for changing mu 2. SOme details

+++ iteration_cycle.py
@@ -9,6 +9,15 @@ import smooth_function
 global sqrt32, sqrt3
 sqrt32 = np.sqrt(3.)/2.
 sqrt3 = np.sqrt(3.)
+global triang_points
+
+def set_num_of_tr_points(number):
+    global triang_points
+    triang_points = number
+
+def get_num_of_tr_points():
+    global triang_points
+    return triang_points
 
 def read_G_imp_frequencies(filename):
     # read file
@@ -65,8 +74,8 @@ def ro():
     # Get the data
     #list
     #[u'count', u'mean', u'tau', u'timeseries', u'error_bins']
-    density_0 = f['simulation/results/density_0/mean/value'].value
-    density_1 = f['simulation/results/density_1/mean/value'].value
+    density_0 = f['simulation/results/density_0/mean/value'][()]
+    density_1 = f['simulation/results/density_1/mean/value'][()]
     print (density_0, density_1)
     charge_density  = density_0 + density_1
     spin_density    = density_0 - density_1
@@ -179,6 +188,48 @@ def calculate_b_vectors(a1, a2, a3):
     
     return b1, b2, b3
 
+def get_b_vectors(lattice_type):
+    if lattice_type == 'square':
+       # real. vectors a1 & a2 & a3
+        a1 = [ 1.0, 0.0 , 0.0]
+        a2 = [ 0.0, 1.0 , 0.0]
+        a3 = [ 0.0, 0.0 , 10.0]
+        # recipr. vectors b1 & b2 & b3
+        b1, b2, b3 = calculate_b_vectors(a1, a2, a3)
+           
+    if lattice_type == 'triangular':
+        # real. vectors a1 & a2 & a3
+#        a1 = [0.866025, -0.5, 0.0]
+#        a2 = [0.866025, 0.5,  0.0]
+        a1 = [ 1.0, 0.0, 0.0]
+        a2 = [ -0.5 , 0.866025 , 0.0]
+        a3 = [0.0, 0.0, 10.0]
+        # recipr. vectors b1 & b2 & b3
+        b1, b2, b3 = calculate_b_vectors(a1, a2, a3)
+    return b1, b2, b3
+
+def get_kstep(lattice_type, Nk):
+    # Nk - number of points
+    # Nk - 1 - number of "slices" of the Mesh
+    b1, b2, b3 = get_b_vectors(lattice_type)
+    if lattice_type == 'square':
+        # compute step in the grid
+        kstep_x = 2.0 * (2. * np.pi) / (Nk - 1)
+        kstep_y = 2.0 * (2. * np.pi) / (Nk - 1)
+        # kstep_x = b1[0] / (Nk - 1)
+        # kstep_y = b2[1] / (Nk - 1)
+        # kstep_z = b3[2] / (Nk - 1)
+        print("kstep_x = " , str(kstep_x), "\nkstep_y = " , str(kstep_y))
+              
+    if lattice_type == 'triangular':
+        # compute step in the grid
+        kstep_x = 2. * (4.0 * np.pi / 3. )/ (Nk - 1)
+        kstep_y = (2. * (2. * np.pi / sqrt3)) / (Nk - 1)
+#        kstep_y = ((2. * np.pi / sqrt3) * 2.) / (Nk - 1)
+#        kstep_x = (2. * (2. * np.pi)) / (Nk - 1)
+        print("kstep_x = " , str(kstep_x), "\nkstep_y = " , str(kstep_y))
+    return kstep_x, kstep_y
+
 def interaction_dispersion(filename, Nk, lattice_type, param):
     
     kpoints = np.array([[[x, y] for x in range(Nk)] for y in range(Nk)]).reshape((Nk * Nk, 2))
@@ -198,12 +249,26 @@ def interaction_dispersion(filename, Nk, lattice_type, param):
     for k in range(Nk * Nk):
         vec_ = [0.0, 0.0]
         for i in range(2):
-            kpoints_coordinates[k][i] = kpoints[k][0] * b1[i]/ (Nk - 1) + kpoints[k][1] * b2[i] / (Nk - 1)
+            if (lattice_type == "square"):
+                kpoints_coordinates[k][i] = kpoints[k][0] * b1[i]/ (Nk - 1) + kpoints[k][1] * b2[i] / (Nk - 1)
+            elif (lattice_type == "triangular"):
+                kpoints_coordinates[k][i] = kpoints[k][0] * b1[i]/ (Nk -1 ) + kpoints[k][1] * b2[i] / (Nk - 1)
     
-    # print(kpoints_coordinates)
+#    print(kpoints_coordinates)
     
     dispersion = open(filename + ".dat", "w")
+    dispersion2 = open("check_dots.dat", "w")
     
+    abs_k = sqrt3
+    abs_b = 4. * np.pi * sqrt3 / 3.
+    
+    x_boundary = -4. * np.pi / 3.
+    y_boundary = -2. * np.pi / sqrt3
+    
+    print("x_boun = {}".format(x_boundary ))
+    print("y_boun = {}".format(y_boundary ))
+    
+    m = 0
     i = 0
     for k in kpoints:
 #        print(k)
@@ -213,19 +278,31 @@ def interaction_dispersion(filename, Nk, lattice_type, param):
         # SQUARE
         if lattice_type == 'square':
             interaction[i] = -2. * param * np.sum(np.cos(-np.pi + k * kstep_x))
-#            interaction[i] = -2. * param * (np.cos(-np.pi + k_x * kstep_x) + np.cos(-np.pi + k_y * kstep_y))
             
         # TRIANG
         if lattice_type == 'triangular':
-            interaction_01 = interaction[0]
-            interaction_02 = interaction[1]
-            interaction_03 = interaction[2]
-                # the first order neighbour
-            interaction[i] = -2. * param[0] * (np.cos(-np.pi + (k_x) * kstep_x) + 2. * np.cos(-np.pi + (0.5 * k_x) * kstep_x) * np.cos(-np.pi + (k_y * sqrt32) * kstep_y))
+            
+            x_coord = x_boundary + k_x * kstep_x
+            y_coord = y_boundary + k_y * kstep_y
+        
+            if  ((x_coord < (-2. * np.pi / 3.)) and y_coord < (-abs_k * x_coord - abs_b) ):
+                interaction[i] += 0.0
+            elif((x_coord < (-2. * np.pi / 3.)) and y_coord > (abs_k * x_coord + abs_b) ):
+                interaction[i] += 0.0
+            elif ((x_coord > (2. * np.pi / 3.)) and y_coord > (-abs_k * x_coord + abs_b) ):
+                interaction[i] += 0.0
+            elif ((x_coord > (2. * np.pi / 3.)) and y_coord < (abs_k * x_coord - abs_b) ) :
+                interaction[i] += 0.0
+            else:
+#                interaction[i] = -2. * param[0] * (np.cos(y_coord) + 2. * np.cos(0.5 * y_coord) * np.cos(x_coord * sqrt32))
+                interaction[i] = -2. * param[0] * (np.cos(x_coord) + 2. * np.cos(0.5 * x_coord) * np.cos(y_coord * sqrt32))
+                m +=1
+            # the first order neighbour
+            #  ???      interaction[i] = -2. * param[0] * (np.cos(x_boundary + (k_x) * kstep_x) + 2. * np.cos(x_boundary + (0.5 * k_x) * kstep_x) * np.cos(y_boundary + (k_y * sqrt32) * kstep_y))
                 # the second order neighbours
-            interaction[i] += -2. * param[1] * (np.cos(-np.pi + (sqrt3 * k_y) * kstep_y) + 2. * np.cos(-np.pi + (1.5 * k_x) * kstep_x) * np.cos(-np.pi + (sqrt32 * k_y) * kstep_y))
+#            interaction[i] += -2. * param[1] * (np.cos(-np.pi + (sqrt3 * k_y) * kstep_y) + 2. * np.cos(-np.pi + (1.5 * k_x) * kstep_x) * np.cos(-np.pi + (sqrt32 * k_y) * kstep_y))
                 # the third order neighbours
-            interaction[i] += -2. * param[2] * (np.cos(-np.pi + (2. * k_x) * kstep_x) + 2. * np.cos(-np.pi + (k_x) * kstep_x) * np.cos(-np.pi + (sqrt3 * k_y) * kstep_y))
+#            interaction[i] += -2. * param[2] * (np.cos(-np.pi + (2. * k_x) * kstep_x) + 2. * np.cos(-np.pi + (k_x) * kstep_x) * np.cos(-np.pi + (sqrt3 * k_y) * kstep_y))
             
         dispersion.write(str(np.round(k_x, 6)))
         dispersion.write('\t')
@@ -233,9 +310,31 @@ def interaction_dispersion(filename, Nk, lattice_type, param):
         dispersion.write('\t')
         dispersion.write(str(np.round(interaction[i],6)))
         dispersion.write('\n')
+        
+        
+        dispersion2.write(str(np.round(x_boundary + k_x * kstep_x, 6)))
+        dispersion2.write('\t')
+        dispersion2.write(str(np.round(y_boundary + k_y * kstep_y,6)))
+        dispersion2.write('\t')
+        
+        if  ((x_coord < (-2. * np.pi / sqrt3)) and y_coord < (-abs_k * x_coord - abs_b) ):
+            dispersion2.write(str(np.round(0.0,6)))
+        elif((x_coord < (-2. * np.pi / sqrt3)) and y_coord > (abs_k * x_coord + abs_b) ):
+            dispersion2.write(str(np.round(0.0,6)))
+        elif ((x_coord > (2. * np.pi / sqrt3)) and y_coord > (-abs_k * x_coord + abs_b) ):
+            dispersion2.write(str(np.round(0.0,6)))
+        elif ((x_coord > (2. * np.pi / sqrt3)) and y_coord < (abs_k * x_coord - abs_b) ) :
+            dispersion2.write(str(np.round(0.0,6)))
+        else:
+            dispersion2.write(str(np.round(interaction[i],6)))
+        dispersion2.write('\n')
+        
         i += 1
         
     dispersion.close()
+    dispersion2.close()
+    
+    set_num_of_tr_points(m)
     
     return interaction, kpoints_coordinates
      
@@ -251,6 +350,7 @@ def new_delta(mu, Nk, t, lattice_type, beta, U):
 
     # init arrays for local quantities
     G_loc       = np.zeros(G_imp_omega.shape, dtype=np.complex64)
+    G_exact       = np.zeros(G_imp_omega.shape, dtype=np.complex64)
     Delta_new   = np.zeros(G_imp_omega.shape, dtype=np.complex64)
    
     # if (Delta.Shape() != G_imp_omega.Shape()):
@@ -266,19 +366,32 @@ def new_delta(mu, Nk, t, lattice_type, beta, U):
 
 #   We don't use + mu in that formula, because mu is implemented in the solver
 #   and you get it as a real part of Sw.dat (Self-energy).
-    
-    for i in range(t_k.__len__()):
-        G_loc += 1.0 / (frequencies + mu - t_k[i] - Sigma)
-    G_loc /= Nk ** 2
+    if (lattice_type == "square"):
+        for i in range(t_k.__len__()):
+            G_loc += 1.0 / (frequencies + mu - t_k[i] - Sigma)
+        G_loc /= Nk ** 2
+    elif (lattice_type == "triangular"):
+        for i in range(get_num_of_tr_points()):
+            G_loc += 1.0 / (frequencies + mu - t_k[i] - Sigma)
+        G_loc /= get_num_of_tr_points()
     
     np.savetxt("G_loc.dat", np.column_stack((frequencies.imag, G_loc.real, G_loc.imag)))
     # G_loc_smooth = smooth_function.smooth_data(frequencies.imag, G_loc, 400, "G_loc_smooth")
+    
+    # Exact solution
+        
+    print("m = {}".format(get_num_of_tr_points()))
+    for i in range(get_num_of_tr_points()):
+        G_exact += 1.0 / (frequencies + mu - t_k[i])
+    G_exact /= get_num_of_tr_points()
+       
+    np.savetxt("G_exact.dat", np.column_stack((frequencies.imag, G_exact.real, G_exact.imag)))
     
     ################################################################################
     #                              New Delta
     ################################################################################
     print ("**Compute new delta function")
-    mix_par = 0.75
+    mix_par = 0.25
     print("Mixing:")
     print('{} of old Delta and {} of new Delta'.format(int((1.0 - mix_par) * 100),  int(mix_par * 100)))
     intermixed_part = (1. / G_loc - 1. / G_imp_omega)
@@ -300,44 +413,6 @@ def corrections(iw, mu, Delta_omega, Sigma_omega, beta, U):
     for i in range(len(g_omega_minus_analytical_part)):
         g_tau_0 += g_omega_minus_analytical_part[i] / beta
     print ('g (tau = 0) =', g_tau_0)
-
-def get_b_vectors(lattice_type):
-    if lattice_type == 'square':
-       # real. vectors a1 & a2 & a3
-        a1 = [ 1.0, 0.0 , 0.0]
-        a2 = [ 0.0, 1.0 , 0.0]
-        a3 = [ 0.0, 0.0 , 10.0]
-        # recipr. vectors b1 & b2 & b3
-        b1, b2, b3 = calculate_b_vectors(a1, a2, a3)
-           
-    if lattice_type == 'triangular':
-        # real. vectors a1 & a2 & a3
-        a1 = [0.866025, -0.5, 0.0]
-        a2 = [0.866025, 0.5,  0.0]
-        a3 = [0.0, 0.0, 10.0]
-        # recipr. vectors b1 & b2 & b3
-        b1, b2, b3 = calculate_b_vectors(a1, a2, a3)
-    return b1, b2, b3
-
-def get_kstep(lattice_type, Nk):
-    # Nk - number of points
-    # Nk - 1 - number of "slices" of the Mesh
-    b1, b2, b3 = get_b_vectors(lattice_type)
-    if lattice_type == 'square':
-        # compute step in the grid
-        kstep_x = 2.0 * (2. * np.pi) / (Nk - 1)
-        kstep_y = 2.0 * (2. * np.pi) / (Nk - 1)
-        # kstep_x = b1[0] / (Nk - 1)
-        # kstep_y = b2[1] / (Nk - 1)
-        # kstep_z = b3[2] / (Nk - 1)
-        print("kstep_x = " , str(kstep_x), "\nkstep_y = " , str(kstep_y))
-              
-    if lattice_type == 'triangular':
-        # compute step in the grid
-        kstep_y = ((2. * np.pi / sqrt3) * 2.) / (Nk - 1)
-        kstep_x = (2. * (2. * np.pi)) / (Nk - 1)
-        print("kstep_x = " , str(kstep_x), "\nkstep_y = " , str(kstep_y))
-    return kstep_x, kstep_y
 
 def new_lambda(beta, interaction, Nk, lattice_type):
     
@@ -379,6 +454,8 @@ def new_lambda(beta, interaction, Nk, lattice_type):
     elif ((Sigma[0] - Sigma[1]) < 0.0):
         metal = False
         print ("Impurity is in the insulator phase")
+    elif (Sigma[0] == Sigma[1]):
+        print( "No interaction. Sigma == 0.0" )
     else:
         print ("WTF with your Sigma? (New Lambda)")
 
@@ -405,7 +482,7 @@ def new_lambda(beta, interaction, Nk, lattice_type):
         (frequencies.imag, Lambda_new_smooth.real, Lambda_new_smooth.imag, Lambda_spin_new.real, Lambda_spin_new.imag)))
 
 
-def rename_files(number_of_iteration):
+def rename_files(number_of_iteration, type_of_calc):
     
     print ("Rename files.")
     
@@ -416,9 +493,10 @@ def rename_files(number_of_iteration):
     # Delta_new.dat -> Delta.dat
     shutil.copy("Delta_new.dat", "Delta.dat")
 
-    # Phi.dat   -> Phi_iteration.dat
-    shutil.copy("Phi.dat", "Phi_" + str(number_of_iteration) + ".dat")
-    os.remove("Phi.dat")
+    if (type_of_calc == "edmft"):
+        # Phi.dat   -> Phi_iteration.dat
+        shutil.copy("Phi.dat", "Phi_" + str(number_of_iteration) + ".dat")
+        os.remove("Phi.dat")
     
-    # Lambda_new.dat + (spin_lambda = 0) -> Phi.dat
-    shutil.copy("Lambda_new.dat", "Phi.dat")
+        # Lambda_new.dat + (spin_lambda = 0) -> Phi.dat
+        shutil.copy("Lambda_new.dat", "Phi.dat")
